+++ runner.go
@@ -1,20 +1,32 @@
 package htindex
 
 import (
-	"archive/zip"
 	"bufio"
 	"fmt"
 	"os"
+	"sync"
 )
 
 func (hti *HTindex) Run() error {
+	fmt.Println(hti)
 	inCh := make(chan string)
 	errCh := make(chan error)
-	go outputError(errCh)
-	go hti.outputInput(inCh, errCh)
+	outCh := make(chan *title)
+	var wg sync.WaitGroup
+	var wgOut sync.WaitGroup
+	wg.Add(hti.jobsNum)
+	wgOut.Add(1)
+	go hti.outputError(errCh)
+	go hti.outputResult(outCh, &wgOut)
+	for i := 0; i < hti.jobsNum; i++ {
+		go hti.worker(inCh, outCh, errCh, &wg)
+	}
 	if err := hti.readInput(inCh, errCh); err != nil {
 		return err
 	}
+	wg.Wait()
+	close(outCh)
+	wgOut.Wait()
 	return nil
 }
 
@@ -34,38 +46,4 @@ func (hti *HTindex) readInput(inCh chan<- string, errCh chan<- error) error {
 		errCh <- err
 	}
 	return nil
-}
-
-func outputError(errCh <-chan error) {
-	for err := range errCh {
-		fmt.Println(err)
-	}
-}
-
-func (hti *HTindex) outputInput(inCh <-chan string, errCh chan<- error) {
-	for file := range inCh {
-		fmt.Println(file)
-		fmt.Println()
-		r, err := zip.OpenReader(hti.rootPrefix + file)
-		if err != nil {
-			errCh <- err
-		}
-		defer r.Close()
-		for _, f := range r.File {
-			fmt.Printf("File: %s\n", f.Name)
-			rc, err := f.Open()
-			if err != nil {
-				errCh <- err
-			}
-
-			scanner := bufio.NewScanner(rc)
-			for scanner.Scan() {
-				fmt.Println(scanner.Text())
-			}
-			if err := scanner.Err(); err != nil {
-				errCh <- err
-			}
-			rc.Close()
-		}
-	}
 }